feat: add extension preview page
